Move HttpUrlConnection implementation to its own package.
