+++ okhttp/src/main/java/com/squareup/okhttp/internal/huc/DelegatingHttpsURLConnection.java
@@ -14,7 +14,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package com.squareup.okhttp.internal.http;
+package com.squareup.okhttp.internal.huc;
 
 import com.squareup.okhttp.Handshake;
 import java.io.IOException;
