+++ okhttp/src/main/java/com/squareup/okhttp/internal/huc/ResponseCacheAdapter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.squareup.okhttp.internal.http;
+package com.squareup.okhttp.internal.huc;
 
 import com.squareup.okhttp.OkResponseCache;
 import com.squareup.okhttp.Request;
@@ -29,7 +29,7 @@ import java.util.List;
 import java.util.Map;
 
 /** Adapts {@link ResponseCache} to {@link OkResponseCache}. */
-public class ResponseCacheAdapter implements OkResponseCache {
+public final class ResponseCacheAdapter implements OkResponseCache {
 
   private final ResponseCache delegate;
